reorganize redux related file structure
